refactor(src): PAGE_SIZE, SORTING_OPTIONS 상수 저장 위치 models/templates으로 변경
