+++ frontend/src/service/getSortingOptionByValue.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_SORTING_OPTION, SORTING_OPTIONS } from '@/api';
+import { DEFAULT_SORTING_OPTION, SORTING_OPTIONS } from '@/models/templates';
 
 export const getSortingOptionByValue = (value: string) =>
   SORTING_OPTIONS.find((el) => el.value === value) || DEFAULT_SORTING_OPTION;
